Create main.py
It is my first project from "Weekly Project", where the main objective is to calculate the yearly, monthly, and weekly total revenue from cupcakes.inc.
My project will show a menu that will ask the user to select an option , on what data to be displayed using a bar chart. (Note: close the graph to continue to another option)
[a] [Yearly revenue.]
[b] [Weekly revenue.]
[c] [Monthly revenue.]
[d] [Custom date interval revenue.]
[e] [Basic cupcakes Revenue vs Delux cupcakes revenue.]
Option [a], [b] and [c], have suboptions, where the user can select the data he wants to see (ex: yearly total revenue, yearly basic cupcakes revenue, yearly deluxe cupcakes revenue)
Option [d], is similar to the previous options (having suboptions ), only that the period will be selected by the user by entering a start date and a stop date. And the revenue will be represented for each day.
Option [e], will also ask the user for a start/stop date and in the bar chart, the revenue from basic cupcakes and delux cupcakes are shown side by side, for each day of the week in the mentioned period. For better comparison.